+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Bitmark Swift SDK v2.6.0 introduces some breaking changes to align with other Bitmark SDKs:
 - Remove `Account.parseAccountNumber(accountNumber:)` in favor of `Account.validateAccountNumber(accountNumber:)`.
 - Remove `AccountNumber.parse()` in favor of `AccountNumber.validate()`
+- `Transfer.respond(withResponseParams:)` will return an optional String as txId when reponding with acceptance.
 
 ## 2.0.0
 ### Features: